+++ packages/components/src/components/segmented-control/segmented-control/segmented-control.tsx
@@ -4,35 +4,38 @@ import {
   Element,
   Event,
   type EventEmitter,
+  forceUpdate,
   Host,
+  h,
   type JSX,
   Listen,
   Prop,
   Watch,
-  forceUpdate,
-  h,
 } from '@stencil/core';
 import type { BreakpointCustomizable, PropTypes, Theme, ValidatorFunction } from '../../../types';
 import {
   AllowedTypes,
-  THEMES,
   attachComponentCss,
+  FORM_STATES,
   hasPropValueChanged,
   observeChildren,
+  THEMES,
   throwIfChildrenAreNotOfKind,
   unobserveChildren,
   validateProps,
   warnIfDeprecatedPropIsUsed,
 } from '../../../utils';
+import { StateMessage } from '../../common/state-message/state-message';
 import type { SegmentedControlItem } from '../segmented-control-item/segmented-control-item';
 import { getComponentCss } from './segmented-control-styles';
 import {
+  getItemMaxWidth,
   SEGMENTED_CONTROL_BACKGROUND_COLORS,
   SEGMENTED_CONTROL_COLUMNS,
   type SegmentedControlBackgroundColor,
   type SegmentedControlColumns,
+  type SegmentedControlState,
   type SegmentedControlUpdateEventDetail,
-  getItemMaxWidth,
   syncSegmentedControlItemsProps,
 } from './segmented-control-utils';
 
@@ -46,11 +49,15 @@ const propTypes: PropTypes<typeof SegmentedControl> = {
   columns: AllowedTypes.breakpoint<SegmentedControlColumns>(SEGMENTED_CONTROL_COLUMNS),
   name: AllowedTypes.string,
   form: AllowedTypes.string,
+  compact: AllowedTypes.boolean,
   disabled: AllowedTypes.boolean,
+  state: AllowedTypes.oneOf<SegmentedControlState>(FORM_STATES),
+  message: AllowedTypes.string,
 };
 
 /**
  * @slot {"name": "", "description": "Default slot for the `p-segmented-control-item` tags." }
+ * @slot {"name": "message", "description": "Shows a state message. Only [phrasing content](https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/Content_categories#Phrasing_content) is allowed."}
  *
  * @controlled { "props": ["value"], "event": "update", "isInternallyMutated": true }
  */
@@ -75,6 +82,15 @@ export class SegmentedControl {
 
   /** The name of the segmented-control. */
   @Prop({ reflect: true }) public name?: string;
+
+  /** A boolean value that, if present, renders the segmented-control as a compact version. */
+  @Prop() public compact?: boolean = false;
+
+  /** Indicates the validation or overall status of the component. */
+  @Prop() public state?: SegmentedControlState = 'none';
+
+  /** Dynamic feedback text for validation or status. */
+  @Prop() public message?: string = '';
 
   /** Sets the amount of columns. */
   @Prop() public columns?: BreakpointCustomizable<SegmentedControlColumns> = 'auto';
@@ -154,12 +170,28 @@ export class SegmentedControl {
     validateProps(this, propTypes);
     warnIfDeprecatedPropIsUsed<typeof SegmentedControl>(this, 'backgroundColor');
 
-    attachComponentCss(this.host, getComponentCss, getItemMaxWidth(this.host), this.columns);
-    syncSegmentedControlItemsProps(this.host, this.value, this.disabled, this.theme);
+    attachComponentCss(
+      this.host,
+      getComponentCss,
+      getItemMaxWidth(this.host, this.compact),
+      this.columns,
+      this.state,
+      this.theme
+    );
+    syncSegmentedControlItemsProps(
+      this.host,
+      this.value,
+      this.disabled,
+      this.state,
+      this.message,
+      this.compact,
+      this.theme
+    );
 
     return (
       <Host role="group" inert={this.disabled}>
         <slot />
+        <StateMessage state={this.state} message={this.message} theme={this.theme} host={this.host} />
       </Host>
     );
   }